% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{kernel_blow}
\alias{kernel_blow}
\title{Generate a proposal for the Blow kernel}
\usage{
kernel_blow(n_dim, p_phi, x, xp)
}
\arguments{
\item{n_dim}{Dimension of the parameter space.}

\item{p_phi}{Probability of updating each coordinate.}

\item{x}{Current point `x`.}

\item{xp}{Current point `x'`.}
}
\value{
A list with the `proposal`, `n_phi`, and `phi`.
}
\description{
Generate a proposal for the Blow kernel
}
\keyword{internal}
