% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{kernel_traverse}
\alias{kernel_traverse}
\title{Generate a proposal for the Traverse kernel}
\usage{
kernel_traverse(n_dim, p_phi, x, xp, beta)
}
\arguments{
\item{n_dim}{Dimension of the parameter space.}

\item{p_phi}{Probability of updating each coordinate.}

\item{x}{Current point `x`.}

\item{xp}{Current point `x'`.}

\item{beta}{Step parameter beta.}
}
\value{
A list with the `proposal` and `n_phi` (number of moved coordinates).
}
\description{
Generate a proposal for the Traverse kernel
}
\keyword{internal}
