% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{kernel_walk}
\alias{kernel_walk}
\title{Generate a proposal for the Walk kernel}
\usage{
kernel_walk(n_dim, p_phi, aw, x, xp)
}
\arguments{
\item{n_dim}{Dimension of the parameter space.}

\item{p_phi}{Probability of updating each coordinate.}

\item{aw}{Scale parameter `a_w` (default 1.5).}

\item{x}{Current point `x`.}

\item{xp}{Current point `x'`.}
}
\value{
A list with the `proposal` and `n_phi`.
}
\description{
Generate a proposal for the Walk kernel
}
\keyword{internal}
