% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{log_density_hop}
\alias{log_density_hop}
\title{Calculate the log of the proposal density for the Hop kernel}
\usage{
log_density_hop(n_phi, phi, h, x, xp)
}
\arguments{
\item{n_phi}{Number of moved coordinates.}

\item{phi}{Boolean vector indicating moved coordinates.}

\item{h}{Proposed point.}

\item{x}{Reference point.}

\item{xp}{Reference point.}
}
\value{
The value of -log(g(h)).
}
\description{
Calculate the log of the proposal density for the Hop kernel
}
\keyword{internal}
