\name{Rtwalk}

\alias{Rtwalk}

\docType{package}

\Rdversion{1.0}

\title{The R implementation of the 't-walk'}

\description{
The 't-walk' is "A General Purpose Sampling Algorithm for Continuous Distributions",
see http://www.cimat.mx/~jac/twalk/ for technical details.
}


\details{

%  \tabular{ll}{
    Package:     Rtwalk \cr
    Type:        Package \cr
    Version:     1.0 \cr
    Date:        2009-10-01 \cr
    License:     GPL (>= 2) \cr
    Copyright:   (c) 2004-2009 J Andr\'es Christen (CIMAT, Guanajuato, MEXICO) \cr
    URL:         \url{http://www.cimat.mx/~jac/twalk/} \cr
%    }


    The t-walk is run with the function:

     \code{Runtwalk}
 
    Please look at the documentation of the  \code{Runtwalk} function for more details, or download the
    file  \url{http://www.cimat.mx/~jac/twalk/examples.R} .


}

\author{
  J Andres Christen (CIMAT, Guanajuato, MEXICO).
}

\seealso{\code{\link{Runtwalk}}}

\keyword{package}
