% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\docType{methods}
\name{twalk-methods}
\alias{twalk-methods}
\alias{print.twalk}
\alias{summary.twalk}
\title{Methods for objects of class 'twalk'}
\usage{
\method{print}{twalk}(x, ...)

\method{summary}{twalk}(object, burnin_frac = 0.2, ...)
}
\description{
S3 methods for inspecting t-walk MCMC output objects.
}
\keyword{internal}
